Feature/wasm trigger

+++ modules/model/src/main/kotlin/jp/co/soramitsu/iroha2/codec/ScaleCodecReader.kt
@@ -98,7 +98,7 @@ class ScaleCodecReader(private val source: ByteArray) {
         }
     }
 
-    fun readUByte(): Short {
+    fun readUByte(): Int {
         return UBYTE.read(this)
     }
